requirements.txt
pyulog==1.2.2
pandas==2.3.1
scikit-learn==1.8.0
matplotlib==3.10.5
seaborn==0.13.2
tabulate==0.10.0
python-dotenv==1.0.1
google-generativeai==0.8.4
openai==1.61.0
anthropic==0.45.2
